refoctor reducers
