+++ .eslintrc.json
@@ -13,10 +13,13 @@
     "ecmaFeatures": {
       "jsx": true
     },
-    "ecmaVersion": 2018,
+    "ecmaVersion": 6,
     "sourceType": "module"
   },
   "plugins": ["react"],
   "parser": "babel-eslint",
-  "rules": {}
+  "rules": {
+    "react/jsx-uses-react": "error",
+    "react/jsx-uses-vars": "error"
+  }
 }
