+++ src/containers/Game/Game.js
@@ -7,7 +7,6 @@ import { setCurrentPlayerAction } from "../../actions/actions.js";
 import { Form } from "react-bootstrap";
 import { Row, Col } from "react-bootstrap";
 
-
 const Game = props => {
   const [playerName, setPlayerName] = useState("Select player");
 
@@ -45,11 +44,11 @@ const Game = props => {
       </Col>
     </Row>
   );
-}
+};
 
 const mapStateToProps = state => {
   return {
-    players: state.manageGameSession.players,
+    players: state.managePlayers.players,
     currentPlayer: state.currentPlayer
   };
 };
@@ -72,51 +71,3 @@ Game.propTypes = {
   currentPlayer: PropTypes.object,
   setCurrentPlayer: PropTypes.func
 };
-
-
-
-//   class Game extends Component {
-//     state = {
-//         playerName: "Select user",
-//         currentPlayer: {},
-//     };
-
-//     generatePlayersList() {
-//         const { players } = this.props;
-//         return players.map((user) => <option key={user.id} value={user.name}>{user.name}</option>)
-//     }
-
-//     changePlayer = (event) => {
-//         this.setState({"playerName": event.target.value});
-//         this.setCurrentPlayer(event.target.value);
-//     }
-
-//     setCurrentPlayer(playerName) {
-//         const arr = [...this.props.players];
-//         const selectedPlayer = arr.filter(user => user.name === playerName);
-//         this.setState({currentPlayer: selectedPlayer});
-//         this.setState({currentPlayerId: selectedPlayer.id});
-//     }
-
-//     renderGameBoard() {
-//         return this.state.playerName !== 'Select user' && <GuessForm currentPlayer={this.state.currentPlayer}/>;
-//     }
-
-//     render() {
-//         return (
-//             <div className="row mh100">
-//                 <div className="col-md-2">
-//                     <h3>Select Player</h3>
-//                     <select className="form-control mt20" onChange={this.changePlayer}>
-//                         <option value='Select user'>Select player</option>
-//                         {this.generatePlayersList()}
-//                     </select>
-//                 </div>
-//                 <div className="col-md-8 offset-md-1">
-//                     { props.currentPlayer.name !== 'Select user' && <GuessForm /> }
-//                     {this.renderGameBoard()}
-//                 </div>
-//             </div>
-//         )
-//     }
-// }
