+++ src/reducers/currentPlayer.js
@@ -1,14 +1,12 @@
 import { SET_CURRENT_PLAYER } from "../actions/actionsList.js";
 
 const intialState = {
-  currentPlayer: {
-    id: 1,
-    name: "Jack",
-    history: {
-      Easy: [],
-      Medium: [],
-      Hard: []
-    }
+  id: 1,
+  name: "Jack",
+  history: {
+    Easy: [],
+    Medium: [],
+    Hard: []
   }
 };
 
@@ -16,7 +14,7 @@ const currentPlayer = (state = intialState, action) => {
   switch (action.type) {
     case SET_CURRENT_PLAYER:
       return {
-        currentPlayer: action.payload
+        ...action.payload
       };
     default:
       return state;
