+++ src/reducers/manageGameSession.js
@@ -1,42 +1,42 @@
-import { UPDATE_HISTORY, CLEAR_HISTORY } from "../actions/actionsList.js";
-import { defaultPlayers } from "../../src/defaultPlayersList.js";
+// import { UPDATE_HISTORY, CLEAR_HISTORY } from "../actions/actionsList.js";
+// import { defaultPlayers } from "../../src/defaultPlayersList.js";
 
-const intialState = {
-  players: defaultPlayers
-};
+// const intialState = {
+//   players: defaultPlayers
+// };
 
-const manageGameSession = (state = intialState, action) => {
-  const playerList = [...state.players];
-  const players = [...state.players];
-  switch (action.type) {
-    case UPDATE_HISTORY:
-      players.forEach(player => {
-        if (player.id === action.payload.id) {
-          player.history[action.payload.activeMode].push({
-            number: action.payload.number,
-            result: action.payload.result
-          });
-        }
-      });
-      return {
-        ...state,
-        players: players
-      };
+// const manageGameSession = (state = intialState, action) => {
+//   const playerList = [...state.players];
+//   const players = [...state.players];
+//   switch (action.type) {
+//     case UPDATE_HISTORY:
+//       players.forEach(player => {
+//         if (player.id === action.payload.id) {
+//           player.history[action.payload.activeMode].push({
+//             number: action.payload.number,
+//             result: action.payload.result
+//           });
+//         }
+//       });
+//       return {
+//         ...state,
+//         players: players
+//       };
 
-    case CLEAR_HISTORY:
-      playerList.forEach(player => {
-        if (player.id === action.payload.id) {
-          player.history[action.payload.mode] = [];
-        }
-      });
-      return {
-        ...state,
-        players: playerList
-      };
+//     case CLEAR_HISTORY:
+//       playerList.forEach(player => {
+//         if (player.id === action.payload.id) {
+//           player.history[action.payload.mode] = [];
+//         }
+//       });
+//       return {
+//         ...state,
+//         players: playerList
+//       };
 
-    default:
-      return state;
-  }
-};
+//     default:
+//       return state;
+//   }
+// };
 
-export default manageGameSession;
+// export default manageGameSession;
