+++ src/reducers/rootReducer.js
@@ -1,11 +1,10 @@
 import { combineReducers } from "redux";
 import managePlayers from "./managePlayers.js";
-import manageGameSession from "./manageGameSession.js";
+// import manageGameSession from "./manageGameSession.js";
 import currentPlayer from "./currentPlayer";
 
 const rootReducer = combineReducers({
   managePlayers,
-  manageGameSession,
   currentPlayer
 });
